# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Vexel Health API (LIMS-first)
  version: 0.1.0
  description: |
    Contract-first OpenAPI spec for Vexel Health Platform.
    All admin endpoints require bearerAuth.
    All requests should include X-Correlation-ID header.
    Tenant is resolved server-side (Host header in prod, x-tenant-id in dev when enabled).
servers:
  - url: /api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Client-supplied correlation ID. If not provided, server generates one.
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [statusCode, message]
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
        correlationId:
          type: string
    Pagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...
    UserSummary:
      type: object
      required: [id, email, tenantId, status]
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        tenantId:
          type: string
        roles:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, disabled, pending]
        isSuperAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Role:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        permissions:
          type: array
          description: List of permission keys assigned to this role
          items:
            type: string
        isSystem:
          type: boolean
          description: True for built-in system roles that cannot be deleted
...
    Patient:
      type: object
      required: [id, tenantId, firstName, lastName, mrn, createdAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
          nullable: true
        ageYears:
          type: integer
          nullable: true
        gender:
          type: string
          nullable: true
        mrn:
          type: string
        mobile:
          type: string
          nullable: true
        cnic:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Encounter:
      type: object
      required: [id, tenantId, patientId, status, createdAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        patientId:
          type: string
        encounterCode:
          type: string
          nullable: true
          description: "LIMS Order ID e.g. VXL-2602-001"
        status:
          type: string
          enum: [registered, lab_ordered, specimen_collected, specimen_received, partial_resulted, resulted, verified, cancelled]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        patient:
          $ref: '#/components/schemas/Patient'
        labOrders:
          type: array
          items:
            $ref: '#/components/schemas/LabOrder'
        specimenItems:
          type: array
          items:
            $ref: '#/components/schemas/SpecimenItem'
    LabOrder:
      type: object
      required: [id, tenantId, encounterId, testId, status, resultStatus, priority, createdAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        encounterId:
          type: string
        testId:
          type: string
        testNameSnapshot:
          type: string
          nullable: true
        test:
          type: object
          description: Resolved test definition (populated when available)
          nullable: true
          properties:
            id:
              type: string
            code:
              type: string
            name:
              type: string
            specimenType:
              type: string
              nullable: true
            loincCode:
              type: string
              nullable: true
        status:
          type: string
          enum: [ordered, specimen_collected, processing, resulted, verified, cancelled]
        resultStatus:
          type: string
          enum: [PENDING, SUBMITTED]
          default: PENDING
        priority:
          type: string
          enum: [routine, stat, urgent]
        submittedAt:
          type: string
          format: date-time
          nullable: true
        submittedById:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        specimen:
          $ref: '#/components/schemas/Specimen'
        results:
          type: array
          items:
            $ref: '#/components/schemas/LabResult'
    Specimen:
      type: object
      required: [id, tenantId, labOrderId, barcode, type, status, createdAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        labOrderId:
          type: string
        barcode:
          type: string
        type:
          type: string
        collectedAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [pending, collected, received, rejected]
        createdAt:
          type: string
          format: date-time
    LabResult:
      type: object
      required: [id, tenantId, labOrderId, value, enteredAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        labOrderId:
          type: string
        parameterId:
          type: string
          nullable: true
        parameterNameSnapshot:
          type: string
          nullable: true
        value:
          type: string
        unit:
          type: string
          nullable: true
        referenceRange:
          type: string
          nullable: true
        flag:
          type: string
          enum: [normal, high, low, critical]
          nullable: true
        locked:
          type: boolean
          default: false
        enteredAt:
          type: string
          format: date-time
        enteredById:
          type: string
          nullable: true
        verifiedAt:
          type: string
          format: date-time
          nullable: true
        verifiedBy:
          type: string
          nullable: true
    SpecimenItem:
      type: object
      required: [id, tenantId, encounterId, catalogSpecimenType, status, createdAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        encounterId:
          type: string
        catalogSpecimenType:
          type: string
          description: "Exact specimen type e.g. 'EDTA Blood', 'Urine'"
        status:
          type: string
          enum: [PENDING, COLLECTED, POSTPONED, RECEIVED]
        barcode:
          type: string
          nullable: true
        collectedAt:
          type: string
          format: date-time
          nullable: true
        collectedById:
          type: string
          nullable: true
        postponedAt:
          type: string
          format: date-time
          nullable: true
        postponeReason:
          type: string
          nullable: true
        receivedAt:
          type: string
          format: date-time
          nullable: true
        receivedById:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    VerificationEncounterSummary:
      type: object
      description: Row item for verification worklist
      required: [encounterId, submittedTestsCount, createdAt]
      properties:
        encounterId:
          type: string
        encounterCode:
          type: string
          nullable: true
        submittedTestsCount:
          type: integer
        oldestSubmittedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        patient:
          $ref: '#/components/schemas/Patient'
    VerificationTestCard:
      type: object
      description: One test with only filled parameters for verification preview
      required: [labOrderId, testName, resultStatus]
      properties:
        labOrderId:
          type: string
        testName:
          type: string
        resultStatus:
          type: string
          enum: [SUBMITTED, VERIFIED]
        submittedAt:
          type: string
          format: date-time
          nullable: true
        filledParameters:
          type: array
          items:
            type: object
            properties:
              parameterId:
                type: string
                nullable: true
              name:
                type: string
              value:
                type: string
              unit:
                type: string
                nullable: true
              referenceRange:
                type: string
                nullable: true
              flag:
                type: string
                enum: [normal, high, low, critical]
                nullable: true
...
paths:
...
  /users:
    get:
      operationId: listUsers
      summary: List users (tenant-scoped)
      tags: [Users]
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, disabled, pending]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      operationId: createUser
      summary: Create a new user (tenant-scoped)
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, firstName, lastName, password]
              properties:
                email:
                  type: string
                  format: email
                firstName:
                  type: string
                lastName:
                  type: string
                password:
                  type: string
                roles:
                  type: array
                  items:
                    type: string
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummary'
        "409":
          description: Email already in use within tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userId}:
    get:
      operationId: getUser
      summary: Get user by ID
      tags: [Users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummary'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      operationId: updateUser
      summary: Update user (name, status)
      tags: [Users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                status:
                  type: string
                  enum: [active, disabled]
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummary'
  /users/{userId}/roles:
    get:
      operationId: getUserRoles
      summary: Get roles assigned to user
      tags: [Users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
    put:
      operationId: setUserRoles
      summary: Set roles for user (replaces current assignment)
      tags: [Users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roleIds]
              properties:
                roleIds:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Updated roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'

  # ─── Roles ────────────────────────────────────────────────────────────────
...
  /encounters/{encounterId}:
    get:
      operationId: getEncounter
      summary: Get encounter by ID
      tags: [Encounters]
      parameters:
        - name: encounterId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Encounter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Encounter'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /encounters/{encounterId}:verify:
    post:
      operationId: verifyEncounter
      summary: Verify lab results (requires result.verify permission)
      tags: [Encounters]
      parameters:
        - name: encounterId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Encounter verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Encounter'
        "409":
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /verification/encounters/pending:
    get:
      operationId: getVerificationQueue
      summary: List encounters with submitted tests pending verification
      tags: [Verification]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        "200":
          description: Verification queue
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/VerificationEncounterSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /verification/encounters/{encounterId}:
    get:
      operationId: getVerificationEncounterDetail
      summary: Get encounter with submitted tests for verification (only filled parameters)
      tags: [Verification]
      parameters:
        - name: encounterId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Encounter verification detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  encounter:
                    $ref: '#/components/schemas/Encounter'
                  patient:
                    $ref: '#/components/schemas/Patient'
                  submittedTestsCount:
                    type: integer
                  pendingVerificationCount:
                    type: integer
                  testCards:
                    type: array
                    items:
                      $ref: '#/components/schemas/VerificationTestCard'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /verification/encounters/{encounterId}:verify:
    post:
      operationId: verifyEncounter
      summary: Verify all submitted tests for an encounter and trigger publish
      tags: [Verification]
      parameters:
        - name: encounterId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: No body required
      responses:
        "200":
          description: Encounter verified; document pipeline enqueued
          content:
            application/json:
              schema:
                type: object
                properties:
                  encounterId:
                    type: string
                  status:
                    type: string
                  documentJobId:
                    type: string
                    nullable: true
        "409":
          description: No submitted tests to verify
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'